Refactor: organize the structure of Layout
레이아웃의 구조 정리 및 스타일 일부 변경
