+++ src/components/Layout/Logo.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const Logo = () => {
+  return <img src="@images/moyaLogo.png" alt="MoYa" className="logo" />;
+};
